requirements.txt
torch
torchvision
mediapipe
pillow
numpy
tqdm
tensorboard